fix: merge counters and compute
